chore: Update main.dart and pubspec.lock
This commit updates the main.dart file to ensure that the loadEnv function is called before any other initialization code. This change ensures that the environment variables are loaded correctly before the app starts.

In addition, the pubspec.lock file is updated to use version 3.6.0 of the archive package and version 6.1.23 of the google_sign_in_android package. These updates address known issues and improve compatibility with other dependencies.

Recent repository commits:
- fix finish
- fix something
- Merge branch 'features/login-with-google'
- Merge branch 'isekai'
- commit to merge new code