feat: Add API service for fetching yoga poses
The code changes in this commit add a new API service called `PoseService` for fetching yoga poses. The `pose_service.dart` file is created, which includes functions for retrieving a list of poses and getting details for a specific pose. These functions make use of the `DioInstance` class for making API requests.

Recent repository commits:
- feat: Add Notification model
- feat: Add DioInstance and blog service for API calls
- commit some model have bug
- commit continue
- commit

```