feat: Add Notification service for fetching notifications
The code changes in this commit add a new API service called `NotificationService` for fetching notifications. The `notification_service.dart` file is imported and included in the `login_page.dart` file. The `onPressed` function in the login button is updated to call the `getNotifications` function from the `NotificationService` instead of the `getExercises` function. This allows the login button to print the notifications instead of the exercises.

Recent repository commits:
- feat: Add API service for fetching yoga poses
- feat: Add Notification model
- feat: Add DioInstance and blog service for API calls
- commit some model have bug
- commit continue
- commit