feat: Refactor patchUserAccount to use PatchUserAccountRequest class
The code changes in this commit refactor the `patchUserAccount` function in the `account_service.dart` file to use the `PatchUserAccountRequest` class instead of a `Map<String, dynamic>`. This class encapsulates the `username` and `phone` properties and provides a `toMap` method for converting the data to a map. This improves code readability and type safety.

Recent repository commits:
- feat: Update login button to use new password change API
- feat: Add account service for fetching user account details
- feat: Add event service for fetching events
- feat: Add Notification service for fetching notifications
- feat: Add API service for fetching yoga poses
- feat: Add Notification model
- feat: Add DioInstance and blog service for API calls
- commit some model have bug
- commit continue
- commit