feat: Add getUserProfile function for fetching user profile
The code changes in this commit add a new function called `getUserProfile` to the `account_service.dart` file. This function makes an API call to retrieve the user's profile information from the server. It uses the `DioInstance` class to send the request and handles the response accordingly. This function will be used to fetch the user's profile data in the application.

Recent repository commits:
- feat: Update login button to use new password change API
- feat: Add account service for fetching user account details
- feat: Add event service for fetching events
- feat: Add Notification service for fetching notifications
- feat: Add API service for fetching yoga poses
- feat: Add Notification model
- feat: Add DioInstance and blog service for API calls
- commit some model have bug
- commit continue
- commit