refactor: Update HomePage route in main.dart
This commit updates the `HomePage` route in the `main.dart` file to pass additional parameters to the `HomePage` widget. The `isVietnamese` parameter is set based on the `_locale` value, and the `isDarkMode` parameter is set based on the `_themeMode` value. The `onThemeChanged` callback is set to `_toggleTheme`, and the `locale` parameter is set to `_locale`. Finally, the `onLanguageChanged` callback is set to `_changeLanguage`. This change enhances the functionality of the `HomePage` widget by providing more customization options.