refactor: Update initialRoute in main.dart
This commit updates the `initialRoute` in the `main.dart` file to use the `AppRoutes.firstScreen` route instead of `AppRoutes.homepage` when the `widget.access` is not null. This change ensures that the `MainScreen` is displayed as the initial screen when the user is already authenticated. The `HomePage` route is still used as the initial screen when the user is not authenticated. This modification improves the user experience by directing authenticated users to the appropriate screen upon app launch.