refactor: Update Activities page to load events from API
This commit modifies the `Activities` page to load events from the API using the `EventService` class. It adds the necessary imports and initializes a list of events in the `_ActivitiesState` class. The `initState` method is updated to call the `_loadEvents` method, which makes an API request to fetch the events and updates the state with the retrieved data. The `itemCount` property in the `GridView.builder` is updated to use the length of the `_events` list, ensuring that the correct number of cards is displayed. The `title`, `caption`, and `remainingDays` properties of the `CustomCard` widget are updated to use the corresponding properties of the events in the `_events` list. This change improves the functionality of the `Activities` page by dynamically loading events from the API.