refactor: Update Blog and BlogDetail pages to use image URL from JSON
This commit updates the `Blog` and `BlogDetail` pages to use the image URL from the JSON data instead of individual properties. In the `Blog` page, the `childAspectRatio` property is changed to 4/5 to maintain the aspect ratio of the images. In the `BlogDetail` page, the `imageUrl` property is used to display the image. This change improves code consistency and allows for more flexibility in handling different image URLs.