refactor: Update BlogVote model to make 'user' property nullable
The code changes in this commit modify the 'BlogVote' model in the 'blog.dart' file. The 'user' property has been changed from a non-nullable string to a nullable string by adding a question mark ('?') after the type. This change allows for cases where the 'user' value may be null.

This commit follows the established convention of using a prefix to indicate the type of change (refactor for code refactoring) and provides a clear and concise description of the changes made.